+++ src/pages/Admin/Food/ListFood.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import { useState } from "react";
 
 import { Space, Table, Input, Button, Image, Popconfirm } from "antd";
 import type { ColumnsType } from "antd/es/table";
@@ -12,6 +12,8 @@ import {
 } from "../../../service/food.service";
 import { IFood } from "../../../interface/model";
 import EditFood from "./EditFood";
+import { RootState } from "../../../store/store";
+import { useAppSelector } from "../../../store/hooks";
 interface DataType {
   id: string;
   name: string;
@@ -24,7 +26,9 @@ const { Search } = Input;
 const ListFood: React.FC = () => {
   const { data: foods } = useFetchFoodQuery();
   const [removeFood] = useRemoveFoodMutation();
-  console.log(foods);
+
+  const { role } = useAppSelector((state: RootState) => state.auth);
+
   const columns: ColumnsType<DataType> = [
     {
       title: "Mã Food",
@@ -52,32 +56,35 @@ const ListFood: React.FC = () => {
       key: "price",
     },
     {
-      render: (_, record) => (
-        <Space size="middle">
-          <EditFood dataFood={record} />
-
-          <Popconfirm
-            placement="topLeft"
-            title="Bạn muốn xóa sản phẩm?"
-            description="Xóa sẽ mất sản phẩm này trong database!"
-            onConfirm={() => removeFood(record.id)}
-            okText="Yes"
-            cancelText="No"
-            okButtonProps={{
-              style: { backgroundColor: "#007bff", color: "white" },
-            }}
-            cancelButtonProps={{
-              style: { backgroundColor: "#dc3545", color: "white" },
-            }}
-          >
-            <Button>
-              <div className="flex ">
-                <DeleteOutlined />
-              </div>
-            </Button>
-          </Popconfirm>
-        </Space>
-      ),
+      render: (_, record) => {
+        if (role === 1) {
+          return (
+            <Space size="middle">
+              <EditFood dataFood={record} />
+              <Popconfirm
+                placement="topLeft"
+                title="Bạn muốn xóa sản phẩm?"
+                description="Xóa sẽ mất sản phẩm này trong database!"
+                onConfirm={() => removeFood(record.id)}
+                okText="Yes"
+                cancelText="No"
+                okButtonProps={{
+                  style: { backgroundColor: "#007bff", color: "white" },
+                }}
+                cancelButtonProps={{
+                  style: { backgroundColor: "#dc3545", color: "white" },
+                }}
+              >
+                <Button>
+                  <div className="flex ">
+                    <DeleteOutlined />
+                  </div>
+                </Button>
+              </Popconfirm>
+            </Space>
+          );
+        }
+      },
     },
   ];
 
@@ -89,13 +96,14 @@ const ListFood: React.FC = () => {
     price: food?.price,
     //   tags: [food.status === 1 ? "Hoạt động" : "Ngừng hoạt động"],
   }));
-  console.log("🚀 ~ file: ListFood.tsx:92 ~ dataFood ~ dataFood:", dataFood)
-  const [dataList, setDataList] = useState<any>(null)
+  const [dataList, setDataList] = useState<any>(null);
 
   const onSearch = (value: any, _e: any) => {
-    const results =dataFood.filter((item: any) => item.name.toLowerCase().includes(value.toLowerCase()))
-    setDataList(results)
-  }
+    const results = dataFood.filter((item: any) =>
+      item.name.toLowerCase().includes(value.toLowerCase())
+    );
+    setDataList(results);
+  };
 
   return (
     <>
@@ -108,14 +116,12 @@ const ListFood: React.FC = () => {
             onSearch={onSearch}
           />
 
-          <AddFood />
+          {role === 1 && <AddFood />}
         </div>
       </div>
       {dataList ? (
         <Table columns={columns} dataSource={dataList} />
-
       ) : (
-
         <Table columns={columns} dataSource={dataFood} />
       )}
     </>
