+++ src/pages/Admin/MovieRoom/ListMovieRoom.tsx
@@ -1,26 +1,23 @@
+import { useState } from "react";
 
-import React, { useState } from "react";
-
-import { Space, Table, Input, Button,  Popconfirm } from "antd";
+import { Space, Table, Input, Button, Popconfirm } from "antd";
 import type { ColumnsType } from "antd/es/table";
 import { DeleteOutlined } from "@ant-design/icons";
-
 
 import {
   useFetchMovieRoomQuery,
   useRemoveMovieRoomMutation,
 } from "../../../service/movieroom.service";
-import {
-useFetchCinemaQuery
-} from "../../../service/brand.service";
+import { useFetchCinemaQuery } from "../../../service/brand.service";
 import { ICinemas, IMovieRoom } from "../../../interface/model";
 import AddMovieRoom from "./AddMovieRoom";
 import EditMovieRoom from "./EditMovieRoom";
+import { RootState } from "../../../store/store";
+import { useAppSelector } from "../../../store/hooks";
 interface DataType {
-
   id: string;
   name: string;
-  id_cinema: string
+  id_cinema: string;
 }
 
 const { Search } = Input;
@@ -28,6 +25,8 @@ const { Search } = Input;
 const ListMovieRoom: React.FC = () => {
   const { data: movies } = useFetchMovieRoomQuery();
   const { data: cinemas } = useFetchCinemaQuery();
+
+  const { role } = useAppSelector((state: RootState) => state.auth);
 
   const [removeMovie] = useRemoveMovieRoomMutation();
   console.log(movies);
@@ -49,35 +48,38 @@ const ListMovieRoom: React.FC = () => {
       dataIndex: "id_cinema",
       key: "id_cinema",
     },
-   
+
     {
+      render: (_, record) => {
+        if (role === 1) {
+          return (
+            <Space size="middle">
+              <EditMovieRoom dataMovieRoom={record} />
 
-      render: (_, record) => (
-        <Space size="middle">
-          <EditMovieRoom dataMovieRoom={record} />
-
-          <Popconfirm
-            placement="topLeft"
-            title="Bạn muốn xóa sản phẩm?"
-            description="Xóa sẽ mất sản phẩm này trong database!"
-            onConfirm={() => removeMovie(record.id)}
-            okText="Yes"
-            cancelText="No"
-            okButtonProps={{
-              style: { backgroundColor: "#007bff", color: "white" },
-            }}
-            cancelButtonProps={{
-              style: { backgroundColor: "#dc3545", color: "white" },
-            }}
-          >
-            <Button>
-              <div className="flex ">
-                <DeleteOutlined />
-              </div>
-            </Button>
-          </Popconfirm>
-        </Space>
-      ),
+              <Popconfirm
+                placement="topLeft"
+                title="Bạn muốn xóa sản phẩm?"
+                description="Xóa sẽ mất sản phẩm này trong database!"
+                onConfirm={() => removeMovie(record.id)}
+                okText="Yes"
+                cancelText="No"
+                okButtonProps={{
+                  style: { backgroundColor: "#007bff", color: "white" },
+                }}
+                cancelButtonProps={{
+                  style: { backgroundColor: "#dc3545", color: "white" },
+                }}
+              >
+                <Button>
+                  <div className="flex ">
+                    <DeleteOutlined />
+                  </div>
+                </Button>
+              </Popconfirm>
+            </Space>
+          );
+        }
+      },
     },
   ];
 
@@ -86,16 +88,19 @@ const ListMovieRoom: React.FC = () => {
       key: index.toString(),
       id: movie.id,
       name: movie?.name,
-      id_cinema: (cinemas as any)?.data?.find((cinemas: ICinemas) =>cinemas.id === movie.id_cinema)?.name
-
+      id_cinema: (cinemas as any)?.data?.find(
+        (cinemas: ICinemas) => cinemas.id === movie.id_cinema
+      )?.name,
     })
   );
-  const [dataList, setDataList] = useState<any>(null)
+  const [dataList, setDataList] = useState<any>(null);
 
   const onSearch = (value: any, _e: any) => {
-    const results =dataMovie.filter((item: any) => item.name.toLowerCase().includes(value.toLowerCase()))
-    setDataList(results)
-  }
+    const results = dataMovie.filter((item: any) =>
+      item.name.toLowerCase().includes(value.toLowerCase())
+    );
+    setDataList(results);
+  };
   return (
     <>
       <div className="">
@@ -107,20 +112,16 @@ const ListMovieRoom: React.FC = () => {
             onSearch={onSearch}
           />
 
-
-          <AddMovieRoom />
+         {role == 1 && <AddMovieRoom />}
         </div>
       </div>
       {dataList ? (
-      <Table columns={columns} dataSource={dataList} />
-        
+        <Table columns={columns} dataSource={dataList} />
       ) : (
-
-      <Table columns={columns} dataSource={dataMovie} />
+        <Table columns={columns} dataSource={dataMovie} />
       )}
     </>
   );
 };
-
 
 export default ListMovieRoom;
